Fix issue where minified code does not work
I have no idea what is causing this to happen but when the Raindrop
class is used as part of fn.toString() inside of a blob/web worker it
gets named the wrong variable.

This is an awful hack and there has to be a better way to handle this.